refactored schema function to separate file
